Little refactoring and try to make it performant
